Share battery compensation; add motor feedforward
Move battery voltage measurement/compensation onto Board (nominal voltage per platform, update_voltage_compensation(), voltage_scale measured at init). DifferentialDrive now reads board.voltage_scale (removed its own redundant voltage logic). EncodedMotor switched to feedforward+P velocity control (kS/kV per platform), holds Board reference, applies voltage_scale to motor effort, and clamps effort to [-1,1]. Minor PID/integral changes to match the new feedforward approach.
